Moved Keyboard Adapter into being its own class
